Move context methods out of the main file
